requirements.txt
asgiref==3.2.10
Django==3.0.6
djangorestframework==3.11.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
mysqlclient==1.4.6
protobuf==3.12.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
